requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
greenlet==0.4.1
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
python-instagram==0.8.0
simplejson==3.3.0
wsgiref==0.1.2
